scoregraph: Fix rounding
In rounding up there was a corner case that some values weren't rounded
up. This happens because of truncation with integers. Convert to
floating point first.

In getDivision() there was another corner case that it was returning 1
when clearly the division points could be further apart. This happened
when span divided by scaleFactor was very small. To fix this convert to
floating point and check that we return at least scaleFactor.

+++ src/components/scoregraph.cpp
@@ -65,14 +65,14 @@ namespace {
 
     inline int roundUp(int value, int factor)
     {
-        return static_cast<int>(ceil(value / factor)) * factor;
+        return static_cast<int>(ceil(static_cast<qreal>(value) / factor)) * factor;
     }
 
     inline int getDivision(int high, int low, int space)
     {
         const int span = high - low;
         const int scaleFactor = pow(10, static_cast<int>(log10(span)) - 1);
-        return std::max(1, span / scaleFactor / space * scaleFactor);
+        return std::max(1, static_cast<int>(static_cast<qreal>(span) / scaleFactor / space)) * scaleFactor;
     }
 } // namespace
